extensive refactoring of dependency handling
All this to try and make all three kinds of dependency
handling work appropriately.

Eventually had to put all kinds of info on the current this to keep
track of dependency so that rendering only gets called after
all dependencies have been satisfied.

A hack, using a title of '__ignore__' delays ending the looping
when getting all the tiddlers in the current context.